+++ src/skulk/extensions/__init__.py
@@ -13,24 +13,31 @@ from skulk.extensions.loader import (
     load_extensions,
     resolve_skulk_version,
 )
+from skulk.extensions.telemetry import ClusterNodeView, snapshot_cluster
 from skulk.extensions.types import (
+    AdvertiseCapability,
     BaseChatMiddleware,
     ChatMiddleware,
     ChatResponseSummary,
     EmbedTexts,
     ExtensionContext,
+    ReadClusterTelemetry,
     SkulkExtension,
 )
 
 __all__ = [
     "ENTRY_POINT_GROUP",
+    "AdvertiseCapability",
     "BaseChatMiddleware",
     "ChatMiddleware",
     "ChatResponseSummary",
+    "ClusterNodeView",
     "EmbedTexts",
     "ExtensionContext",
     "LoadedExtensions",
+    "ReadClusterTelemetry",
     "SkulkExtension",
     "load_extensions",
     "resolve_skulk_version",
+    "snapshot_cluster",
 ]